backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
python-multipart==0.0.10
python-dotenv==1.0.1
asyncio==3.4.3
pydantic==2.10.3 